Merged PR 2207: Attempt to implement databricks notebooks via asset bundles
This is the deploying the `databricks-notebooks` directory to databricks. Note that the path to the dev is different than the path to the other environments. This is because development path needs to reference `current_user` otherwise it errors out.

This is the dev deployment bundle location:
![Screenshot 2024-12-03 at 14.44.55.png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/2207/attachments/Screenshot%202024-12-03%20at%2014.44.55.png)

For all other deployments, this will be `/Workspace/adf-pipelines/{env}/files/...`.

I have tested it on the dev moneyview pipeline:
![Screenshot 2024-12-04 at 10.22.27.png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/2207/attachments/Screenshot%202024-12-04%20at%2010.22.27.png)

During review, could you pay attention to the devops pipelines order, please?
Also the naming of the jobs, I'm not super sure about it.

PS:
This PR also adds a PR template.