Merged PR 5239: CDP/Update the pipeline for mocean and cbs datasets
## Why this PR is needed:
This PR introduces alternative to the `pl_rest_api_cloud` pipeline for the CBS datasets. The pipeline allows to download an arbitrary number of files (as opposed to 2) and creates 1 singular output file in the data lake. The API requires to have a top and skip options available to select the next number of lines to query.

This pipeline should replace `pl_rest_api_cloud`, however, it is not possible to get the same results for the mocean dataset which uses the same pipeline. Hence this new pipeline was added.

## Related Tickets & Documents

- link to Jira issue to be solved by this PR:
  - https://jira/browse/CDP-5810
- related PR's in other repositories:
   - todo: add triggers update

## Example outcome of this change

![image.png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/5239/attachments/image.png)

The inside of the loop looks like this:
![image (2).png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/5239/attachments/image%20%282%29.png)

## Instructions on how to verify quality of this PR

I have run this pipeline on the `cbs_kwb_2024_85984NED`. and `cbs_kwb_2024_85984NED_nested`

- [cbs pipeline run](https://adf.azure.com/en/monitoring/pipelineruns/933b2884-bad9-40dc-8540-aa6e1c40498a?factory=%2Fsubscriptions%2Fd12cd7ca-ec96-4566-8890-89f3194d1169%2FresourceGroups%2Frg-ingestion-cdp-preprd-we-ojsu%2Fproviders%2FMicrosoft.DataFactory%2Ffactories%2Fadf-ing-cdp-preprd-dev-we-ojsu)

The run was validated using the row count numbers in databricks using the following queries - results pasted as comments:
![image (5).png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/5239/attachments/image%20%285%29.png)

- [cbs nested pipeline run](https://adf.azure.com/en/monitoring/pipelineruns/c415b142-c53d-4c98-86b3-ac62bb045b36?factory=%2Fsubscriptions%2Fd12cd7ca-ec96-4566-8890-89f3194d1169%2FresourceGroups%2Frg-ingestion-cdp-preprd-we-ojsu%2Fproviders%2FMicrosoft.DataFactory%2Ffactories%2Fadf-ing-cdp-preprd-dev-we-ojsu)

![image (3).png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/5239/attachments/image%20%283%29.png)
![image (4).png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/5239/attachments/image%20%284%29.png)

Which are different results but after using the compiled code from DBT, where each of the columns is unpacked - we do not see any null IDs
![image (6).png](https://dev.azure.com/NIBC/59b9b987-65c8-4616-8698-a2240ca2c182/_apis/git/repositories/ac1561d1-06fd-42d0-a2cf-f5cb21c85910/pullRequests/5239/attachments/image%20%286%29.png)

## [optional] Are there a...