Merged PR 5312: Fix the typing in CBS pipeline again.
<!--
#csf:3.3.5h:We use a PR template to remind everyone of our way of working and the Cloud Security Framework.
-->
## Why this PR is needed:

This pipeline works in debug mode but when trying to pass the scheduled runs, there is an error with the value type.

This could fix:
```
Operation on target Set lines to skip failed: The function 'mul' expects its first parameter to be an integer or a decimal number. The provided value is of type 'String'.
```