Merged PR 5584: CDP-6011/Adding RFC related section to the pull request template
<!--
#csf:3.3.5h:We use a PR template to remind everyone of our way of working and the Cloud Security Framework.
-->
## Why this PR is needed:

This PR updates the pull request template. This change is meant to help with the check of resources which are released to production. It introduces a section in the Pull Request template which marks the production change.

This change also requires all developers to be consistent with filling in the PR templates fully so there is actually a point in this. It would be the responsibility of the approvers to make sure the PR template is filled in fully.

## 🚦 Does this PR affect production?
<!-- Production related changes could be for example releasing a new dataset to production,
updating a model or documentation of a model which already has a model maturity set to prod. -->
- [ ] This change **affects production**
- [X] This change **does not affect production**

## Related Tickets & Documents

- link to Jira issue to be solved by this PR:
  - https://jira/browse/CDP-6011
- related PR's in other repositories:
  - https://dev.azure.com/NIBC/Data%20Platform/_git/dbt-cdp/pullrequest/5518

## I (creator of this PR):

- [X] ✨ have chosen a clear title for this PR
  - This will greatly help my colleagues understand where changes come from
  - Template for PR name: starting with the related Jira issue and actively describing what this change does, e.g. `CDP-2024/Add PR Template`
- [X] 👷‍♀️ created a small PR covering a single issue, allowing for effective review.
- [X] 🙌 am aware of and followed best practices documented on [the wiki](https://dev.azure.com/NIBC/Data%20Platform/_wiki/wikis/Data-Platform.wiki/1525/Azure-Data-Factory)